add stats to scheduler
