+++ airflow/settings.py
@@ -116,9 +116,6 @@ except:
 
 configure_logging()
 Stats = configure_stats()
-Stats.gauge('configured', 1)
-log.info("Stats.gauge('configured', 1)")
-
 configure_vars()
 configure_orm()
