+++ tests/utils/stats/test_stackdriverlogger.py
@@ -4,15 +4,15 @@ from mock import patch, mock, MagicMock
 from google.cloud import monitoring
 from google.cloud.monitoring import MetricKind, ValueType
 
-from airflow.stats_config import StackdriverLogger
+from airflow.utils.stats.stackdriverlogger import StackdriverLogger
 
 
-class TestStatsSettings(unittest.TestCase):
+class TestStackdriverLogger(unittest.TestCase):
     @patch('google.cloud.monitoring.Client')
     def test_StackdriverLogger(self, mock_client):
         prefix = 'pre'
 
-        logger = StackdriverLogger(mock_client, prefix, publish=False)
+        logger = StackdriverLogger(mock_client, prefix, start_publishing=False)
         logger.incr("ctr", 5)
         logger.decr("ctr", 2)
         logger.gauge("gauge", 4.2)
